refactor(backend): drop the remaining MySQL/Dolt traces
The app has been running on SQLite for a while, but the persistence layer
was still written against MySQL: sqlc generated MySQL code, and both repos
classified driver errors as *mysql.MySQLError, which SQLite never returns.

- switch the sqlc engine to sqlite and regenerate both sqlgen packages
- replace the shopmap :copyfrom queries, unsupported by the sqlite engine,
  with per-row :exec inserts inside the existing transaction
- schema: int UNSIGNED -> integer, so sqlc types the column as int64
- pkg/mymysql -> pkg/mysqlite, classifying sqlite3.Error constraint codes
- rename dolt.go to sqlite.go / gorm.go by repo style, drop config/dolt

Duplicate logins now actually report myerr.ErrAlreadyExists: besides the
error type, user/repo ran gormDB.AutoMigrate before InitUsers, so the users
table got created without UNIQUE(login) and the conflict never happened.
InitUsers alone owns the table now; repo.User stays only as the association
target for the GORM repos.

+++ internal/backend/shopmap/repo/sqlgen/models.go
@@ -18,7 +18,7 @@ type ShopMap struct {
 
 type ShopMapCategory struct {
 	MapID    string
-	Number   uint32
+	Number   int64
 	Category string
 }
